pawmi-ml/requirements.txt
# Core ML
pandas>=2.2.0
numpy>=2.1.0
scikit-learn>=1.5.0

# Model Training
pycaret>=3.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Utilities
loguru>=0.7.0
pyyaml>=6.0.0
jsonschema>=4.20.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Optional: Synthetic data generation
# sdv>=1.9.0  # Uncomment if using CTGAN
